requirements.txt
PyMuPDF==1.21.1
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
django-storages==1.12.3
Pillow==8.4.0
django-nested-admin==3.3.3
boto3==1.20.16
requests==2.28.2
djangorestframework==3.14.0
